feat(monkey): added comments and floating point number support

+++ evaluator/builtins.go
@@ -13,9 +13,9 @@ var builtins = map[string]*object.Builtin{
 			}
 			switch arg := args[0].(type) {
 			case *object.Array:
-				return &object.Number{Value: int64(len(arg.Elements))}
+				return &object.Number{Value: float64(len(arg.Elements))}
 			case *object.String:
-				return &object.Number{Value: int64(len(arg.Value))}
+				return &object.Number{Value: float64(len(arg.Value))}
 			default:
 				return makeError("argument to `len` not supported, got %s", args[0].Type())
 			}
